Move files to different folders to correct project structure
