+++ frontend/document_collaboration_system/src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./css/App.css";
-import DocumentEditor from "./documentEditor";
+import DocumentEditor from "./documents/documentEditor";
 import { useHistory } from "react-router-dom";
 
 function App({ document, setDocument }) {
