+++ frontend/document_collaboration_system/src/comments/commentWidgets.js
@@ -2,8 +2,8 @@ import React, { useState } from "react";
 import Button from "@material-ui/core/Button";
 import EditCommentDialog from "./editCommentsDialog";
 import CommentsDialog from "./commentsDialog";
-import "./css/base.css";
-import api from "./services/APIService";
+import "../css/base.css";
+import api from "../services/APIService";
 
 export function CommentButton({ document }) {
   const [openCommentsList, setOpenCommentsList] = useState(false);
