+++ frontend/document_collaboration_system/src/comments/placeCommentDialog.js
@@ -6,8 +6,8 @@ import DialogActions from "@material-ui/core/DialogActions";
 import DialogContent from "@material-ui/core/DialogContent";
 import DialogContentText from "@material-ui/core/DialogContentText";
 import DialogTitle from "@material-ui/core/DialogTitle";
-import { AppContext } from "./";
-import api from "./services/APIService";
+import { AppContext } from "../";
+import api from "../services/APIService";
 
 export default function FormDialog({
   openModalWindow,
