+++ frontend/document_collaboration_system/src/common/header.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./css/base.css";
+import "../css/base.css";
 
 function Header() {
   return (
