+++ frontend/document_collaboration_system/src/common/navbar.js
@@ -1,8 +1,8 @@
 import React, { useContext, useEffect } from "react";
-import "./css/base.css";
+import "../css/base.css";
 import { Link } from "react-router-dom";
 import { useHistory } from "react-router-dom";
-import { AppContext } from "./";
+import { AppContext } from "../";
 
 function Navbar() {
   const history = useHistory();
