+++ frontend/document_collaboration_system/src/documents/list.js
@@ -4,8 +4,8 @@ import List from "@material-ui/core/List";
 import Divider from "@material-ui/core/Divider";
 import { useHistory } from "react-router-dom";
 import DocumentListItem from "./documentListItem";
-import api from "./services/APIService";
-import { AppContext } from "./";
+import api from "../services/APIService";
+import { AppContext } from "../";
 
 const useStyles = makeStyles((theme) => ({
   root: {
