+++ frontend/document_collaboration_system/src/invites/inviteUserDialog.js
@@ -4,8 +4,8 @@ import DialogContent from "@material-ui/core/DialogContent";
 import { TextField } from "@material-ui/core";
 import DialogActions from "@material-ui/core/DialogActions";
 import Button from "@material-ui/core/Button";
-import api from "./services/APIService";
-import { AppContext } from "./";
+import api from "../services/APIService";
+import { AppContext } from "../";
 
 export default function InviteUserDialog({ openDialog, setOpen, document }) {
   const [username, setUsername] = useState("");
